[#147] 모듈 구조 정리 및 함수 위치 변경
