+++ rupring/src/core/bootings/mod.rs
@@ -1,0 +1,1 @@
+pub mod aws_lambda;
