+++ rupring/src/core/mod.rs
@@ -1,5 +1,7 @@
 mod banner;
 pub mod boot;
+mod bootings;
+mod compression;
 mod graceful;
 mod parse;
 
@@ -31,71 +33,6 @@ use crate::header::preprocess_headers;
 use crate::logger::print_system_log;
 use crate::swagger::context::SwaggerContext;
 use crate::IModule;
-
-pub fn handle_graceful_shutdown(
-    application_properties: &application_properties::ApplicationProperties,
-    service_avaliable: Arc<AtomicBool>,
-    running_task_count: Arc<AtomicU64>,
-) {
-    let signal_flags = graceful::SignalFlags::new();
-    let shutdown_timeout_duration = application_properties.server.shutdown_timeout_duration();
-
-    if let Err(error) = signal_flags.register_hooks() {
-        print_system_log(
-            Level::Error,
-            format!("Error registering signal hooks: {:?}", error).as_str(),
-        );
-    } else {
-        print_system_log(Level::Info, "Graceful shutdown enabled");
-
-        let service_avaliable = Arc::clone(&service_avaliable);
-        let running_task_count = Arc::clone(&running_task_count);
-        tokio::spawn(async move {
-            let sigterm = Arc::clone(&signal_flags.sigterm);
-            let sigint = Arc::clone(&signal_flags.sigint);
-
-            loop {
-                if sigterm.load(std::sync::atomic::Ordering::Relaxed) {
-                    print_system_log(
-                        Level::Info,
-                        "SIGTERM received. Try to shutdown gracefully...",
-                    );
-                    service_avaliable.store(false, std::sync::atomic::Ordering::Release);
-                    break;
-                }
-
-                if sigint.load(std::sync::atomic::Ordering::Relaxed) {
-                    print_system_log(
-                        Level::Info,
-                        "SIGINT received. Try to shutdown gracefully...",
-                    );
-                    service_avaliable.store(false, std::sync::atomic::Ordering::Release);
-                    break;
-                }
-
-                tokio::time::sleep(tokio::time::Duration::from_millis(500)).await;
-            }
-
-            let shutdown_request_time = std::time::Instant::now();
-
-            loop {
-                if running_task_count.load(std::sync::atomic::Ordering::Relaxed) == 0 {
-                    print_system_log(Level::Info, "All tasks are done. Shutting down...");
-                    std::process::exit(0);
-                }
-
-                // timeout 지나면 강제로 종료
-                let now = std::time::Instant::now();
-                if now.duration_since(shutdown_request_time) >= shutdown_timeout_duration {
-                    print_system_log(Level::Info, "Shutdown timeout reached. Forcing shutdown...");
-                    std::process::exit(0);
-                }
-
-                tokio::time::sleep(tokio::time::Duration::from_millis(500)).await;
-            }
-        });
-    }
-}
 
 pub async fn run_server(
     application_properties: application_properties::ApplicationProperties,
@@ -131,7 +68,7 @@ pub async fn run_server(
     let running_task_count = Arc::new(AtomicU64::new(0));
 
     if is_graceful_shutdown {
-        handle_graceful_shutdown(
+        graceful::handle_graceful_shutdown(
             &application_properties,
             Arc::clone(&service_avaliable),
             Arc::clone(&running_task_count),
@@ -406,7 +343,7 @@ fn post_process_response(
     match application_properties.server.compression.algorithm {
         CompressionAlgorithm::Gzip => {
             // compression
-            let compressed_bytes = compress_with_gzip(&response.body);
+            let compressed_bytes = compression::compress_with_gzip(&response.body);
 
             let compressed_bytes = match compressed_bytes {
                 Ok(compressed_bytes) => compressed_bytes,
@@ -430,7 +367,7 @@ fn post_process_response(
         }
         CompressionAlgorithm::Deflate => {
             // compression
-            let compressed_bytes = compress_with_deflate(&response.body);
+            let compressed_bytes = compression::compress_with_deflate(&response.body);
 
             let compressed_bytes = match compressed_bytes {
                 Ok(compressed_bytes) => compressed_bytes,
@@ -456,25 +393,4 @@ fn post_process_response(
     }
 
     response
-}
-
-fn compress_with_gzip(body: &[u8]) -> anyhow::Result<Vec<u8>> {
-    use std::io::Write;
-
-    let mut encoder = flate2::write::GzEncoder::new(Vec::new(), flate2::Compression::default());
-    encoder.write_all(body)?;
-    let compressed = encoder.finish()?;
-
-    Ok(compressed)
-}
-
-fn compress_with_deflate(body: &[u8]) -> anyhow::Result<Vec<u8>> {
-    use std::io::Write;
-
-    let mut encoder =
-        flate2::write::DeflateEncoder::new(Vec::new(), flate2::Compression::default());
-    encoder.write_all(body)?;
-    let compressed = encoder.finish()?;
-
-    Ok(compressed)
 }
